[ATC] Refine some concepts and some code.
We need to expand the notion of statement term with a notion of a 'loop flag'
that determines whether the statement term occurs at the end of a computation
path in a loop body or not. This way, we can more precisely check that recursive
calls to the loop functions occur exactly at the end of all the computation
paths.