[ATC] Optimize symbolic execution.
Change the definition of ADD-VAR (one of the functions used for the symbolic
execution states) slightly to be more similar to CREATE-VAR, to simplify some
rules.

Avoid pushing CREATE-VAR into layers of symbolic execution states, instead
replacing it immediately with ADD-VAR, given a hypothesis that the variable does
not already exist in the scope, which is now more efficiently and simply handled
via a newly introduced predicate and associated rules.