[ATC] Simplify and optimize symbolic execution.
Introduce a new function UPDATE-VAL that replaces WRITE-VAR (under the condition
that the latter does not return an error) in the canonical representation of
states. This UPDATE-VAL function "caches" (i.e. unconditionally satisfies)
certain properties of WRITE-VAR that would otherwise have to repeatedly
established during symbolic execution. This UPDATE-VAR function is now used in
the canonical representation of symbolic computation states instead of
WRITE-VAR. As part of this change, several rules in the symbolic execution are
now unconditional, or have fewer conditions that are more quickly discharged.