[ATC] Improve some generated theorems.
These are the correctness theorems for the non-recursive target functions.

Use more readable bindings in the theorem's formula instead of generating terms
with those substitutions already applied.

Also fix a bug in which ATC was generating terms of the form (BINARY-+ ARG),
i.e. with one argument instead of two: the problem was previously masked by the
fact that untranslation was turning that into a valid (+ ARG).