[Yul] Improve DeadCodeEliminator proof.
The fact that the executable counterpart of some nullary fixtype constructor was
enabled required its disabling in the proof (otherwise the proof would fail).
However, we know disable the executable counterparts of the nullary constructors
of the MODE fixtype when that fixtype is defined: as discussed in the added
documentation about that, this seems appropriate. Thus, we no longer need to
explicitly disable (:E MODE-LEAVE) in the DeadCodeEliminator proof, which seemed
awkward and a bit out of the blue.