[ATC] Extend the symbolic execution theory for arrays.
This adds rules for the extended symbolic computation states. The new rules
handle array updates and their interaction with the various functions that build
the symbolic computation states. Array updates are pushed into the innermost
computation state term, and are normalized according to the syntactic ordering
on their pointers, similarly to how we treat variable updates in C loop proofs.

Array updates may happen in proofs of both C functions and C loops: in the first
case, the pointers are ACL2 variables; in the second case, the pointers are
(READ-VAR ...) terms. Thus, we have rules to separately order the pointers in
these two cases.

More rules to discharge hypotheses need to be added to the symbolic execution
theory. All or most of those rules should already exist in the development.