[ATC] Extend generated correctness theorems.
Also generate hypotheses that the pointers are not null. This derives from the
fact that the ACL2 representation of the C code manipulates arrays directly (as
structures containing lists of values), therefore implicitly implying that there
are non-null pointers to those arrays.

Indeed, some of the recently added rules for symbolic execution related to
arrays have non-null pointer hypotheses.