[ATC] Rephrase a predicate in the symbolic execution.
The previous definition of WRITE-ARRAY-OKP exhibited an unfortunated interaction
with guards involving equalities about array lengths, which occur in some
idiomatic examples. The problem is that ACL2 may use the equality to replace a
term with another one inside a larger term, causing the larger term to no longer
have the expected form for which the symbolic execution has rules, thus making
the proof fail.

The new definition of WRITE-ARRAY-OKP avoids this issue.