[ATC] Improve generated variables.
Use ADD-SUFFIX-TO-FN instead of ADD-SUFFIX to generate pointer variables in
generated theorems, because if the starting variable happens to be in the
COMMON-LISP package and not imported into the ACL2 package, an error occurs.
This way we ensure that the generated pointer variable names are put in the ACL2
package instead of the COMMON-LISP package (if the starting symbol is in the
COMMON-LISP package).