[ATC] Extend ATC's input validation.
Ensure that every recursive target function, which represents a C loop, is
called by a later target function in the list passed to ATC. This recursively
ensures that all the C loops will appear in some C function. It also ensures,
together with the pre-existing check that there is at least one target function,
that we always generate at least one C function.