[ATC] Improve and fix the C static semantics.
The improvement refers to adding the check that the body of a function
definition is a compound statement (the abstract syntax allows any statement).
This was not really a problem because ATC only generates function definitions
whose bodies are compound statements.

The fix refers to not creating a new empty scope for the block that is the
function body, nested into the scope that contains the function parameters: the
function parameters belong to the scope of the body. This could have caused some
illegal C programs to be accepted and generated by ATC (if they re-declare a
formal parameter in the outermost block of the function body), but those would
have failed compilation, and the absence of regressions means that we did not
generate any program like that with ATC.