[ATC] Add a check during code generation.
We must check that an expression term returns a non-void type. The check applies
to the case in which the expression term is a call of a non-recursive target
ACL2 function, which could otherwise return void.

This turns, in an ATC example created by Eric Smith, an internal implementation
error in ATC into a proper error: the ATC code was expecting a certain situation
to never happen, but the expectation was violated because the check added by
this commit was missing.

The example in question is meant to be legal, but that will be addressed in a
separate commit.