[ATC] Improve calculation of affected variables.
This calculation is carried out on the body of non-recursive target functions.
The updated code now takes into account the possibility that the body calls a
preceding (recursive or non-recursive) target function, so we take the latter
function's affected variables as the ones affected by (that part of) the term.