[Yul] Simplify dynamic semantics.
Put the function state outside of the computation state, since it does not get
updated in the same way as the global and local state: it is an environment that
only needs to be passed as an input to the ACL2 execution functions, without
needing to be returned as output. (The addition of the functions in a block to
the function state is realized via a separate ACL2 function, which is called to
augment the function state.)

Adapt correctness proof of the DeadCodeEliminator transformation. This proof is
now simpler; the only "drawback" is that it no longer exhibits some of the proof
approaches needed for more complex transformations, namely the transformation of
computation states (because now we only need to transform the function state,
and not the computation state any more), but that more general approach is
clear, and will be used in upcoming proofs of other transformations.