[ATJ] Add a :NO-AIJ-TYPES option.
This serves to check that the generated Java code does not use any of the AIJ
types, i.e. that it looks more like idiomatic Java, as opposd to mimicking ACL2
in Java.

These checks are carried out during input processing, which may not be as
appropriate a place to do this as after some of the pre-translation steps that
infer types for everything (so that there is more information available).
Nevertheless, it looks like this may be doing all the right checks. However, it
has not been tested thoroughly, and may need revisions, including moving the
checks to a later stage of the code generation process.