[ATC] Add a rule to the symbolic execution.
This splits BOOLEANP over IFs. It is similar to analogous rules for predicates
for C values: even though BOOLEANP is not a predicate for C values, it is used
in a similar fashion on results of EXEC-TEST.

Thanks to Karthik Nukala for running ATC on an example that failed and thus
revealed the need for this rule addition.