[Yul] Improve formalization of variable renaming.
When adding a new renaming pair to the omap, also check that the new variable is
not already in the map's values, besides checking that the old variable is not
already in the map's keys. This is needed for proving dynamic correctness. It
may also let us simplify the static correctness proof by just taking variable
renaming into consideration, and not variable uniqueness.

This added check is justified by the fact that the new code must be statically
safe: if the check fails, it means that the new code is not statically safe, and
so that run of the transformation would be incorrect anyways, and no proof for
that run could be generated.