[Std/util] Extend DEFMAPPING.
Generate disabled theorems by default, and add support for a :THM-ENABLE option
to enable some or all of them.

This aligns with a general philosophy of being more conservative in enabling
rules, to make proofs generally more robust and maintainable, requiring users to
be more explicit about what they want enabled.

This extension also applied to DEFISO, DEFSURJ, and DEFINJ, which are derived
from DEFMAPPING.