[ATJ] Improve location of the generated files.
If a Java package name is specified, now the generated Java files are written in
a subdirectory of :OUTPUT-DIR that corresponds to the package name, according to
common Java convention for organizing source files. See the updated user doc for
details.

Updated trust tags as needed, due to the use of OSLIB::MKDIR to create the
subdirectories where the Java files go.