[ATC] Add a pretty-printing option.
When this option is set, conditional expressions immediately nested in
conditional expressions are parenthesized. This is not necessary according to
C's precedence rules, but it may make the code more readable in the presence of
nested conditional expressions.

This option has been added by adding an extensible collection of pretty-printing
options, where the collection for now only includes this one option. However, it
is envisioned that this will be extended to control more aspects of the
pretty-printing, such as indentation, line lengths, etc.

This particular option is realized by (when the option is set) artifically
lowering the required grade of the 'then' and 'else' subexpressions of
conditional expressions, to the grade just below conditional expressions: this
forces the parenthesization of conditional expressions that appear there in
order to meet the lowered expected grade, while leaving all other kinds of
expressions unchanged in that respect.

Thanks to Karthik Nukala and Eric Smith for using ATC to generate C code with a
deeply nested conditional expression, and for asking whether ATC could put
additional parentheses. This led to this addition, and to the extensible design
of pretty-printing options.