Fixed bug in handling of stobjs in DO loop$ expressions.  Updated *acl2-exports*.
The do loop$ bug fix is explained starting with this comment in
books/projects/apply/loop-tests.lisp, which is above one of a few
additional tests there.

; Until around mid-January 2022 there was a bug in the handling of stobjs by DO
; loop$s, due to lack of sufficient attention to single-threadedness.  The
; assert-event below the following definition had failed; it now succeeds.

The bug fix actually involved some major reworking of the algorithm
for compiling DO loop$ expressions.  It's actually simpler now (e.g.,
one pass instead of two), but there is a new restriction on DO loop$
expressions, disallowing let bindings of WITH-bound variables.  See
for example bad-superior-let-binding in the loop-tests book mentioned
above.  Comments have been updated (as one would expect).  One change
is elimination of the do-body-tuple record, which I think would cause
a lot of books to require recertification.  So while I was at it, I
finished a long-pending move of some symbols in
*acl2-exports-exclusions*, book books/misc/check-acl2-exports.lisp,
into *acl2-exports*.