[APT] Fix bug in ISODATA.
Some code to generate lemma instances of hints contained incorrect logic that
could lead to a call of the built-in FORMALS on NIL in certain cases.

Thanks to Eric Smith for running ISODATA on an example that resulted in an error
from FORMALS being called on NIL, whose investigation led to the discovery of
the bug.