[APT] Extend the user doc of ISODATA.
Mention the circumstances in which the FORTH mapping is pushed into the IF
structure of the old function's body. This pushing was done a while ago, but the
user documentation was not adequately updated at the time. Actually, that change
is somewhat questionable in retrospect, and may be reverted, or made available
as an option (possibly a compatibility option); in any case, at least now the
documentation and the implementation are consistent in that respect.