[APT] Improve ISODATA.
Generate new forms of the new-to-old and old-to-new theorems, just locally for
now, that are the new desired forms for those theorems.

This is a step in that direction, making sure the new theorems can be proved
from the old ones in all our uses of ISODATA.

The next step is to make the current ones local and the new ones the exported
ones.