Added built-in utility er-soft, allowing some specific error types to be suppressed much like warning$.  Fixed build issue with recently-added xdoc/init book.  Tweaked sources so that "make proofs" again works.
Quoting :doc note-8-5:

  A new [event], [set-inhibit-er-soft], allows the user to turn off
  error output of various types.  The related utility
  [toggle-inhibit-er-soft] can turn on or off a single type of error
  output.  Non-[local] versions of these utilities are
  [set-inhibit-er-soft!] and [toggle-inhibit-er-soft!].  Many error
  messages cannot yet be controlled this way, but this may be
  remedied somewhat with community feedback.  Thanks to Eric Smith
  for a request to inhibit step-limit error output, which led to this
  enhancement.

Modified books/xdoc/init.lisp to avoid dependency issue for
regressions.

Regarding the tweak to the sources so that "make proofs" again works:
it was to introduce set-ld-history-entry-user-data with defproxy
before using defstub later in the build.

Passed "make proofs" and "make devel-check" in addition to the usual
"make regression-everything".  Also built ACL2(p) and ran a few tests.