[ATC] Add a missing check.
In the ACL2 representation of C code, ATC was accepting array write patterns
regardless of the presence of the C::DECLAR and C::ASSIGN wrappers, contrary to
what the ATC user documentation states. A check has been added to ensure that
no wrappers are present in the array write pattern.

Thanks to Eric Smith for finding the issue in a test and for reporting it.