[ATC] Fix bugs in struct support.
Some of the checks to recognize the representation of struct operations had to
be relaxed to take into account the package names of the symbols introduced by
DEFSTRUCT, which may be in any package.

It may be still possible, though unlikely, that the checks may fail in the
presence of packages that import symbols generated by DEFSTRUCT from other
packages. This will be addressed eventually, with more robust checks.