[ATC] Improve handling of structs.
Carry through a table of the structure types, which is always a subset of the
DEFSTRUCT table in the world, which starts empty, and eventually contains all
the information for the targeed structure types.

This lets us more easily enforce proper ordering, and makes it easier to access
information for proof generation.