[ATC] Improve DEFSTRUCT.
Store information in the DEFSTRUCT table by the SYMBOL-NAME of the tag symbol,
instead of the tag symbol. This will let us retrieve the information more
flexibly, in particular from a C type, which only has the name.

(The table was already disallowing different keys with the same SYMBOL-NAME, so
this does not cause conflicts.)