[ATC] Extend DEFSTRUCT.
Have it generate theorems saying that if something is a structure of that type,
then it is not an error.

These theorems parallel similar ones for arrays, but while the ones for arrays
are fixed, the ones for stuctures depend on the structure definitions. Thus, the
names of these theorems are stored in the DEFSTRUCT table.