[RAC] Small parser improvements
Fix a potential read out of bounds in main when the filename is
larger that 80 characters.

Now, when the parsing fails, we abort without trying to generate
code. This was leading to some segfault and strange behavior (like
empty files generated).

Remove debugs flags and enable all optimizations. Also refactor
Makefile for faster build.