[ATC] Rephrase fixtype def of values.
Use a tagged sum instead of a flat sum.

This is needed so that we can extend this type with arrays and structs, which
require mutual recursion, currently not supported by DEFFLATSUM.

Aside from DEFFLATSUM not supporting mutual recursion (which could be remedied),
the use of DEFTAGSUM for values is also a step towards more differentiation and
independence between the deep and shallow embeddings of C in ACL2.