Disallow "//" in filenames.
Quoting :doc note-8-5:

  ACL2 will now generally signal an error if a filename contains
  consecutive directory separators, i.e., "//" (in non-Windows
  systems).

While some Lisps (including CCL and SBCL, but not GCL) accept "//" in
a filename, apparently converting it to "/", the presence of "//" in
an include-book argument can confuse the build system.  It seems best
to disallow "//".  There may still be places in the ACL2 sources that
let this by, but probably very few since system function our-truename
now does this check and it is probably called in most cases when a
filename is presented to ACL2.