[ATC] Generalize model of the heap.
Now the heap (which is actually a model of "outside storage" more generally, and
there are plans to rename it accordingly) can contain any value, not only arrays
and structures. This paves the way to pass by pointer, and side-effect, not only
arrays and structures, but also scalar values, which is a way to handle "output
parameter" in C.

The previous notion of 'object', defined as either an array or a struct, has
been removed. Note that that was a temporary notion, not even terminologically
accurate: in C, an object is any region of memory that contains (bits that can
be interpreted as) a value, not just memory regions with arrays and structures.