Fixed bug in raw Lisp code pertaining to types, for stobj fields of stobjs.  Fixed a couple of stack overflows for Allegro CL.
The bug, introduced after Version 8.4, showed up in Allegro CL (but
not in CCL or SBCL; probably I've run successfully in LispWorks too,
not sure).  Allegro CL is relatively slow; maybe it does more error
checking at safety 0.  Anyhow, the error was in using a stobj name
as a Lisp type.  (I know better; it was a dumb coding error.)