[C] Add generic array read/write operations.
These work on any array, unlike the functions specific to the integer
arrays. The deep embedding will be eventually changed to use these new
operations instead of the ones specific to the integer arrays, which really
belong just to the shallow embedding; at that time, the symbolic execution will
be extended to rewrite these generic operations to the ones specific to integer
arrays, under appropriate assumptions that will be satisfied for the generated C
code.