[ATC] Generalize and simplify C dynamic semantics.
For the execution of assignments to array elements, use the generic array
operations instead of the integer-array-type-specific operations with a case for
each. As a result, now all array types are supported, and the code is simpler.

Adapt the proofs of the symbolic execution rules for assignments to array
elements. Now that there is a less direct connection (as intended) between array
assignments in the deep embedding and array writes in the shallow embedding, the
proofs of these rules need to "bridge" a larger gap.