[ATC] Improve DEFSTRUCT.
In order to accommodate array members properly, the current definition of the
generated struct recognizers needs to be generalized so that it uses the
recognizers of the shallowly embedded C types, rather than just the types of the
values. This is because an array value does not have, in its own definition, the
invariant that all its elements have the element type (it would be awkward to
incorporate this invariant in the definition of (array and other) values,
because it would necessitate simultaneously defining TYPE-OF-VALUE.

So this commit changes the definition, and also needs to adapt the generated
proofs for the struct member readers and writers.