[ATC] Improve input validation.
When determining the types of formal parameters, check struct predicates against
the preceding target structs, not against the DEFSTRUCT table, consistently with
what the user doc in fact says.

Thanks to Karthik Nukala for running into an ATC error message that was not
particularly informative and that led to this improvement.