[ATC] Add placeholders in execution.
These are for reading and writing of elements of array members of structures.
Given some previous simplifications and streamlining of the dynamic semantics,
it is now feasible to add somewhat ad hoc support for these operations, so that
ATC can generate proofs for reads and writes of elements of array members of
structures sooner. It is still planned to generalize the dynamic semantics to a
more uniform and comprehensive model where expressions return values or object
designators, but that requires some slightly deeper adaptation of some of the
symbolic execution rules.