[ATC] Extend dynamic semantics.
Add support for execution of assignments to elements of array members of
structures. Similarly to the addes support for reading elements of array members
of structures in the previous commit, this is also a somewhat ad hoc extension,
aimed at adding proof generation for writes to elements of array members of
structures sooner.