[FTY] Improve DEFRESULT.
Avoid hardwiring a stack of function symbols into the structure of errors. Let
errors be anything, and have the current macros and binders construct a more
general stack in which function symbols and other information are more flexibly
mixed. This stack construction will be actually made an option, and not the
default, in future commits.