[ATC] Slightly restrict dynamic semantics.
Require all the expressions used in assignments, except for assignments to
variables, to be pure. This is consistent with ATC's requirements, but somehow
the dynamic semantics was a bit more liberal, allowing the right expression of
an assignemnt to be non-pure. This could actually be problematic in other
appliations (not ATC), because the order of execution of subexpressions is
underspecified in C, while the dynamic semantics was picking a specific order.