[C] [ATC] Add array initializers.
Extend abstract syntax with explicit notion of initializer, which may be a
single expression or a list of expressions, where the latter represents an array
initializer.

Adapt abstract syntax operations, pretty-printer, static semnantics, dynamic
semantics, code generation, and proof generation to the change.

For now array initializers are actually prohibited, but this is just the first
step towards adding support for them.