Summary: Add and improve some :doc and change "connect to" [directory] to "change to" in installation instructions.  Clarify forcing wording when linear arithmetic is used on an equality.  Tweak acl2-doc code to avoid byte compiler warning.  Fork emacs/ into books/emacs/.  Reduce noise from memoize failures and print better message for ACL2 stobj primitives that are inlined as macros in raw Lisp.  Improved efficiency of len.  Changed :doc topic name, prover-output => output-controls.  Avoid putting 'global-stobjs property when it's nil.  Give better output when :pe is applied to built-ins like IF and LET that aren't normal primitives (which were already handled).
Quoting :doc note-8-5:

  Files from the emacs/ directory in the ACL2 distribution have been
  copied to directory books/emacs/.  The original emacs/ directory
  will continue to be maintained for Emacs 24 (though those who have
  used it comfortably with more recent versions may wish to continue
  to load files from the emacs/ directory).  The books/emacs/
  directory is intended for use with recent versions of ACL2, and the
  ACL2 community is welcome to modify it.  Thanks to Warren Hunt for
  noting that users of recent Emacs versions could experience
  problems using files maintained for Emacs 24.

  Failed attempts to [memoize] or [profile] now generally result in
  less noisy error output.  Also, a more useful error message is
  printed when attempting to [memoize] or [profile] a function that
  is introduced either by [defabsstobj] or by [defstobj] with keyword
  argument :inline t, and hence is a macro in raw Lisp; thanks to
  Shilpi Goel for suggesting this improvement.

  The function [len] is significantly more efficient than before.
  (Also, its raw Lisp implementation is simpler, though as before it
  causes an error for extremely (impossibly?) long lists (in Lisp
  parlance, whose length is not a fixnum).)  Thanks to Stephen
  Westfold for encouraging this change and suggesting a suitable
  declaim form, and thanks to Eric Smith for providing helpful tests.

Below are additional details on some of the changes.

Added :doc backquote (and pointers to it, :doc comma and :doc
comma-atsign).  Thanks to Eric Smith for the suggestion.

Clarified in :doc print-object$ and :doc print-object$+ that that
output channel must be an object output channel.  Thanks to Eric Smith
for the suggestion.

Thanks to Warren Hunt for a query that led to the improved forcing wording.

Thanks to Warren Hunt for raising the issue of byte compiler warnings
for emacs code, which led to the acl2-doc fix and also to forking
emacs/ into books/emacs/, as noted above from :DOC note-8-5.  The fork
of emacs/ into books/emacs/ is explained in their respective README
files.  I modified :DOC emacs to explain about which directory to use.
Here is the explanation in emacs/README.

....
NOTE: This directory is maintained for Emacs 24.  If you are using
more recent versions of Emacs, consider using files in books/emacs/
instead (i.e., relative to the present directory, ../books/emacs/).
....

I have run meta-x byte-compile-file on all four .el files in ./emacs/
using Emacs 24.5.1 and on the corresponding ./books/emacs/ files using
Emacs 28.1, both on a 2019 Macbook Pro; there were no warnings.

I encourage the community to modify those four files in books/emacs/
to keep them working with recent Emacs versions.  My current intention
is that the top-level emacs/ directory will continue to be maintained
for Emacs 24, though it may work in later versions (presumably some
have been already using this directory for awhile with later
versions).

By the way, in case anyone is wondering if we really need two emacs/
directories: yes, we do.  There was a big change to tags after Version
24, and I'm pretty sure I'm not the only one who hates that change and
therefore wants to stick with Emacs 24 for as long as possible.  I
don't want to do any more messing with tags myself, but the community
is welcome (as far as I'm concerned) to modify files in books/emacs/,
to implement something more than I've already done in emacs-acl2.el
(in particular for meta-,), or different from it, to make Emacs 28
behave more like Emacs 24 with respect to tags.

Thanks to Eric Smith for suggesting the :pe improvement.