Change xdoc text rendering of &times; to \times.  Add comment about &nbsp; handling in xdoc.
Rendering &times; as x was sometimes confusing.  Regarding the
comment: here it is, as added to parse-xml.lisp.

     (:NBSP  . " ") ; wart: disappears when preceded only by spaces except in <code>..</code>

I spent a little time looking into how to fix this &nbsp; problem.  It
might not be so easy, as I've explained in a comment added to the
definition of function tokens-to-terminal in file
books/xdoc/display.lisp.