[ATC] Extend dynamic semantics.
Handle assignments to (elements of) arrays stored in static storage.  In our
current C subset and formalization, if the variable contains an array value, we
return a pointer to the variable instead: this way, the rest of the dynamic
semantics of assignments to array elements, which expects pointers, will operate
correctly.

Adjust the execution rules during symbolic execution to the change.