[ABNF] Avoid an unnecessary local event.
This may have been needed in the past, but it is no longer needed. In fact,
interestingly, the presence of this local causes an LD failure of this file,
even if it certifies fine. Not clear whether this is related to the similar kind
of behavior mentioned in the previous commit. (Also investigating this issue
further.)