[C] Improve the definition of unary minus.
Define the dynamic semantics of unary minus solely in terms of the deep
embedding, not of the shallow embedding.

Add a more general ACL2 function to calculate the C integer result of an integer
operation from the ACL2 integer (i.e. exact) result of the operation. This is
used to define unary minus, but it will also be used for upcoming reformulations
of other integer operations. The general documentation of signed vs. unsigned
arithmetic has been moved from :DOC INTEGER-OPERATIONS to this new ACL2
function, where it fits better.

Introduce an alternative definition of unary minus as its old (i.e. previous)
defintition. This requires a proof of course, but keeps the proofs of the
symbolic execution rules for unary minus essentially unchanged. However, this
alternative definition proof takes a bit of time due to case explosion, so we
will eliminate it soon, "moving" its proof to the proofs of the symbolic
execution rules.