[C] Improve formulation of dynamic semantics.
Add an operation to test (scalar) values, and eliminate the EXEC-TEST function.

This required changes to some symbolic execution rules and/or their proofs. It
also required an extension of the loop test correctness theorem generated by
ATC, as explained in the documentation of the ATC code that generates that
theorem.