[cert.pl] Attempt to fix cert.pl.
Recent changes broke cert.pl for many of us.  The issue seems to be that it isn't finding the system books dir using the ACL2 environment variable, due to the function determine_acl2_dirs returning before it can reach the fourth way of setting acl2_books.  So I just conditionalized that return (in the code for the third way).

Big disclaimer: I don't know perl, or anything about cert.pl, so somebody should review this.