[ATC] Add two missing validation checks.
As stated in the user doc, declaration of local variables of pointer type is not
supported. However, ATC was failing to check that the type of declared local
variables is not a pointer. This commit adds the missing check.

Thanks to Eric Smith for running into this issue, which actually manifested as a
failure in the proofs generated by ATC, but now correctly fails validation.