[C] Refactor some code and add a theorem.
Factor out predicate to check if a value is a struct with a flexible array
member, since we'll need this for new symbolic execution rules.

Also add a theorem saying that the removal of the flexible array member is a
no-op when the aforementioned predicate does not hold. This will be also used in
symbolic execution.