[ATC] Add a validation check.
This is already included in the user doc, but was not implemented yet. The check
is that structures with flexible array members can be only passed by pointer to
functions, not by value. Passing by value is allowed in C (which drops the
flexible array member), but in our shallow embedding recognized by ATC we
prohibit that.