[ATC] Generalize generation to files.
Instead of generating a translation unit that gets pretty-printed to a (file
system) file (on disk), now ATC generates a C file (intended as an abstract
syntactic entity, which has been recently introduced), that is pretty-printed to
a (file system) file (on disk). Adapt the generated events, including the
theorems.

This provides the right organization for generating code in multiple files.
This organization should not change in the short and medium term.