[ATJ] Extend the no-AIJ_types mode.
Support the use of ACL2 characters and their translation to Java character
literals.

Support the use of ACL2 strings and their translation to Java string
expressions.

Thanks to Eric Smith and Karthik Nukala for working on an application that will
benefit from this new capability, and for suggesting this extension based on
that.