[ABNF] Add a new `defgrammar` tool.
This defines (i.e. builds) an ACL2 representation of an ABNF grammar from a
file, by parsing the file (using the verified grammar parser) and abstracting
the resulting concrete syntax tree to an abstract syntax tree, which is assigned
to a named constant whose name is specified by the user.

Thia macro also provides options to keep the constant closed in output, and to
generate theorems stating certain properties of the grammar.

This macro also provides XDOC integration.

Additional events can be added after ///, as with other macros.