[Std/system] Add a `genvar$` utility.
This is a logic-mode guard-verified version of the built-in `genvar`, which
seems to require a bit of work to put into logic mode.

This new utility uses `magic-ev-fncall` to call `genvar`, and thus it needs a
`state` argument, which `genvar` does not need. However, the upside is that it
can be used in "statically type-checked" code.