Fixed bug in pick-a-point-subset-hint.
This came up when certifying the book
books/projects/rp-rewriter/lib/mult3/medw-compress-correct.lisp in
LispWorks, specifically, the proof of
medw-compress-c-arg-lst-aux-correct.  Apparently the CAR was being
taken of the symbol RP::SIGN-MATTERS.  The bug was actually in the
function harvest-trigger, which was taking the caar of a clause when
the car of that clause was that non-nil symbol.  Inspection of the
book with the buggy definition of harvest-trigger,
books/std/osets/computed-hints.lisp, had another similar bug, so I
fixed that too (both noted by "Matt K." comments).

..... debugging notes .....

Here I'll record how I re-created and debugged the problem using CCL
with safety 3.  By the way, the reason I called this (on the first
line above) a "bug in pick-a-point-subset-hint" is that the second
form on the stack was a call of that function.

cd books/projects/rp-rewriter/lib/mult3/
acl2
(ubt 1)
(set-compiler-enabled nil state)
(value :q)
(declaim (OPTIMIZE (SAFETY 3)))
(lp)
(ld "medw-compress-correct.port")
(set-debugger-enable t)
(in-package "RP")
(rebuild "medw-compress-correct.lisp" 'medw-compress-c-arg-lst-aux-aux-correct-corol)
(defret medw-compress-c-arg-lst-aux-correct
  <... etc.>; see medw-compress-correct.lisp)
:b ; for backtrace
(:form 0)
(pprint *) ; shows offending call of computed-hints::harvest-trigger