[ATC] Eliminate some symbolic execution rules.
These were splitting rules that distribute specific functions over `if`s, which
may lead to case splits, which in general slow down proofs. The `if`
distributivity rules left appear to be needed, because attempting to remove them
causes some of our ATC applications to fail.

All of this may be superseded once the new modular proof generation approach is
in place, but for now this provides an easy and strict improvement.