[ATC] Improve some data representation.
Use an alist for the instantiation computed by a function, instead of a doublet
list. This was we can readily use an existing alist type, without having to
"duplicate" that for doublet lists. The alist is converted to a doublet list
where needed.

This also caught a minor type error in which an alist from symbols to
pseudo-terms was not quite that, due to uses of unquoted string constants.
Everything worked fine because the slight guard violation was in a program-mode
function, and the called function was producing the right thing anyways. This
was actually the one spot that needs the stronger type for the instantiation, in
order to put more code in logic mode.