[ATC] Put remaining functions in logic mode.
Now just one program-mode function remains in `generation.lisp`.  This function
calls `fmt1!` indirectly to actually print the C code to file, so there seems to
be little room for putting this into logic mode (except for using
`magic-ev-fncall`. But it is a small function, so static type checking would not
provide much benefit.

Also, it is run not during `make-event` expansion, but when the events are
processed: ATC does not immediately generate the file, but rather it generates
an event that generates the file. Actually, the fact that this program-mode
function runs during event processing instead of during expansion is probably of
little importance.