Improve with-supporters so that set-state-ok isn't needed.
If you are using with-supporters, then presumably you want the
definitions it generates without needing to assert that they are
allowed to take state.  Thanks to Eric Smith for pointing to
books/kestrel/helpers/advice-code-only.lisp as an example where that
had been necessary.  I fixed with-supporters and also removed
(set-state-ok t) from the book above.  (Note that :DOC note-8-6-books
already says about with-supporters just that it "has been
substantially enhanced", so there is no new release note.)